Move encrytion/decryption functions to external.utils.
